MONEYAPP/requirements.txt
matplotlib==3.4.3
pandas==1.3.3
Pillow==8.3.2
numpy==1.21.2
